small improvement

+++ script/bulk_import/uploads_importer.rb
@@ -41,23 +41,25 @@ module BulkImport
         result_set.close
       end
 
-      surplus_upload_ids = output_existing_ids - source_existing_ids
+      if (surplus_upload_ids = output_existing_ids - source_existing_ids).any?
+        if @settings[:delete_surplus_uploads]
+          puts "Deleting #{surplus_upload_ids.size} uploads from output database..."
 
-      if @settings[:delete_surplus_uploads]
-        puts "Deleting #{surplus_upload_ids.size} uploads from output database..."
-
-        surplus_upload_ids.each_slice(TRANSACTION_SIZE) do |ids|
-          placeholders = (["?"] * ids.size).join(",")
-          @output_db.execute(<<~SQL, ids)
+          surplus_upload_ids.each_slice(TRANSACTION_SIZE) do |ids|
+            placeholders = (["?"] * ids.size).join(",")
+            @output_db.execute(<<~SQL, ids)
             DELETE FROM uploads
             WHERE id IN (#{placeholders})
           SQL
+          end
+
+          output_existing_ids -= surplus_upload_ids
+        else
+          puts "Found #{surplus_upload_ids.size} surplus uploads in output database. " \
+                 "Run with `delete_surplus_uploads: true` to delete them."
         end
 
-        output_existing_ids -= surplus_upload_ids
-      else
-        puts "Found #{surplus_upload_ids.size} surplus uploads in output database. " \
-               "Run with `delete_surplus_uploads: true` to delete them."
+        surplus_upload_ids = nil
       end
 
       max_count = (source_existing_ids - output_existing_ids).size
